名前変更&開催1日前のイベントを取得するAPI作成
