+++ routes/web.php
@@ -19,7 +19,7 @@ $app->get('/api', function(Request $request){
 	return JsonResponse::create($data,200,[]); 
 });
 
-use App\MyLibs\API;
+use App\MyLibs\Series;
 use App\MyLibs\Redis;
 use App\MyLibs\Events;
 
@@ -27,11 +27,17 @@ $app->get('/test/new',function(){
 	$key = 'cache_events';
 
 	$client = new Redis();
-	$res = API::getEvents();
+	$res = Series::getEvents();
+	return JsonResponse::create($res);
 	$events = new Events( $client->get( $key ) );
 	$diff = $events->diff( new Events( $res ) );
 	if( count( $diff ) !== 0 ) $client->set($key, $res);
-	return JsonResponse::create($diff);
+	return JsonResponse::create($res);
+});
+
+$app->get('/test/previous',function(){
+	$res = Series::getOpenPreviousday();
+	return JsonResponse::create($res);
 });
 
 $app->get('/tweet', function() use($app)
